Simplify exclusion bit range module and type

+++ src/cdomain/value/cdomains/int/bitfieldDomain.ml
@@ -660,7 +660,7 @@ module BitfieldFunctor (Ints_t : IntOps.IntOps): Bitfield_SOverflow with type in
 
   let refine_with_bitfield ik x y = meet ik x y
 
-  let refine_with_excl_list ik t (excl : (int_t list * (int64 * int64)) option) : t = norm ik t
+  let refine_with_excl_list ik t (excl : (int_t list * (int * int)) option) : t = norm ik t
 
   let refine_with_incl_list ik t (incl : (int_t list) option) : t =
     let joined =match incl with
